requirements.txt
python-telegram-bot
dotenv
psycopg2
python-dotenv
gTTS